+++ sdk/servicebus/azure-servicebus/tests/async_tests/test_queues_async.py
@@ -69,7 +69,17 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
             with pytest.raises(ServiceBusConnectionError):
                 await (sb_client.get_queue_receiver(servicebus_queue.name, session_id="test", max_wait_time=5))._open_with_retry()
 
+            with pytest.raises(ValueError):
+                sb_client.get_queue_receiver(servicebus_queue.name, max_wait_time=0)
+
             async with sb_client.get_queue_receiver(servicebus_queue.name, max_wait_time=5) as receiver:
+
+                with pytest.raises(ValueError):
+                    await receiver.receive_messages(max_wait_time=0)
+
+                with pytest.raises(ValueError):
+                    await receiver.get_streaming_message_iter(max_wait_time=0)
+
                 count = 0
                 async for message in receiver:
                     print_message(_logger, message)
@@ -730,7 +740,7 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                         print("Finished second sleep", message.locked_until_utc, utc_now())
                         assert message._lock_expired
                         try:
-                            await message.complete()
+                            await receiver.complete_message(message)
                             raise AssertionError("Didn't raise MessageLockExpired")
                         except MessageLockExpired as e:
                             assert isinstance(e.inner_exception, AutoLockRenewTimeout)
@@ -739,12 +749,12 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                             print("Remaining messages", message.locked_until_utc, utc_now())
                             assert message._lock_expired
                             with pytest.raises(MessageLockExpired):
-                                await message.complete()
+                                await receiver.complete_message(message)
                         else:
                             assert message.delivery_count >= 1
                             print("Remaining messages", message.locked_until_utc, utc_now())
                             messages.append(message)
-                            await message.complete()
+                            await receiver.complete_message(message)
             await renewer.close()
             assert len(messages) == 11