+++ sdk/servicebus/azure-servicebus/tests/test_subscriptions.py
@@ -46,11 +46,25 @@ class ServiceBusSubscriptionTests(AzureMgmtTestCase):
                 message = ServiceBusMessage(b"Sample topic message")
                 sender.send_messages(message)
 
+            with pytest.raises(ValueError):
+                sb_client.get_subscription_receiver(
+                    topic_name=servicebus_topic.name,
+                    subscription_name=servicebus_subscription.name,
+                    max_wait_time=0
+                )
+
             with sb_client.get_subscription_receiver(
                     topic_name=servicebus_topic.name,
                     subscription_name=servicebus_subscription.name,
                     max_wait_time=5
             ) as receiver:
+
+                with pytest.raises(ValueError):
+                    receiver.receive_messages(max_wait_time=-1)
+
+                with pytest.raises(ValueError):
+                    receiver.get_streaming_message_iter(max_wait_time=0)
+
                 count = 0
                 for message in receiver:
                     count += 1